071521 Esteban Organized DB & API files, set up server & api router
