+++ api/users.js
@@ -1,0 +1,4 @@
+const express         = require("express");
+const usersRouter     = express.Router();
+
+module.exports = usersRouter;
