+++ package.json
@@ -18,7 +18,7 @@
     "axios": "^0.19.2",
     "express": "^4.17.1",
     "morgan": "^1.10.0",
-    "pg": "^8.2.1",
+    "pg": "^8.6.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
